Refactor and restructure MCP code
